+++ R/RcppExports.R
@@ -9,8 +9,8 @@ fstmetadata <- function(fileName) {
     .Call(`_fst_fstmetadata`, fileName)
 }
 
-fstretrieve <- function(fileName, columnSelection, startRow, endRow) {
-    .Call(`_fst_fstretrieve`, fileName, columnSelection, startRow, endRow)
+fstretrieve <- function(fileName, columnSelection, startRow, endRow, mergeGap = NULL) {
+    .Call(`_fst_fstretrieve`, fileName, columnSelection, startRow, endRow, mergeGap)
 }
 
 fstappend <- function(fileName, table, compression, uniformEncoding) {